requirements.txt
aiogram==3.7.0
APScheduler==3.10.4
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
aiosqlite==0.20.0
pydantic==2.7.1
